Refactor GUI rendering out of the main renderer
